com.github.jackalsin.leetcode.problems/src/test/java/_1351_1400/_1372_Longest_ZigZag_Path_in_a_Binary_Tree/SolutionITest.java
package _1351_1400._1372_Longest_ZigZag_Path_in_a_Binary_Tree;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;
import utils.binaryTree.TreeNode;
import utils.binaryTree.TreeNodes;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author jacka
 * @version 1.0 on 4/22/2020
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase1(Solution solution) {
    final TreeNode root = TreeNodes.getTreeLevelOrder(1, null, 2, 3, 4, null, null, 5, 6, null, 7, null, null, null,
        8, null, 9);
    assertEquals(3, solution.longestZigZag(root));
  }

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase2(Solution solution) {
    final TreeNode root = TreeNodes.getTreeLevelOrder(1, 1, 1, null, 1, null, null, 1, 1, null, 1);
    assertEquals(4, solution.longestZigZag(root));
  }

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase3(Solution solution) {
    final TreeNode root = TreeNodes.getTreeLevelOrder(1);
    assertEquals(0, solution.longestZigZag(root));
  }

  static Stream<Solution> solutionStream() {
    return Stream.of(
        new SolutionI()
    );
  }
}